Reseolved Warning: findDOMNode is deprecated in StrictMode.
This bug depended on animations effects.
When I set transition to false in OverlayTrigger all of animations was removed with this warnig. So i decided to add this missing class to Popover.
Result is in half satisfactory :
Warnig is GONE
Fade effect is ON during hover offect
But Fade offect after hover effects is GONE
If someone will find class responsible for this effect let me know.